backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.11.0
httpx>=0.24.0,<0.25.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
supabase==2.15.3
python-multipart>=0.0.6
